###### Required for all non-trivial PRs
- [x] Rebased/mergable
- [x] Tests pass
- [ ] CHANGELOG.md updated

`Deduplicate` is called from various places in the engine and can cause
a lot of garbage to get created.  It first creates a map and then
adds each value to the map in order (1st alloc).  It then creates a
new slice (2nd alloc) and appends everything from the map to the slice.
Finally, it sorted the new slice (3rd alloc).

This switches the algorithm to use stable sorting and reusing the existing
slice to avoid allocations.

```
$ benchcmp /tmp/old.txt /tmp/new.txt
benchmark                         old ns/op     new ns/op     delta
BenchmarkValues_Deduplicate-8     447726        57155         -87.23%

benchmark                         old allocs     new allocs     delta
BenchmarkValues_Deduplicate-8     67             1              -98.51%

benchmark                         old bytes     new bytes     delta
BenchmarkValues_Deduplicate-8     140583        32            -99.98%
```